fix(examples): inject react runtime for ui package
