+++ examples/kitchen-sink/packages/ui/package.json
@@ -12,8 +12,8 @@
   ],
   "scripts": {
     "test": "jest",
-    "build": "tsup src/index.tsx --format esm,cjs --dts --external react",
-    "dev": "tsup src/index.tsx --format esm,cjs --watch --dts --external react",
+    "build": "tsup src/index.tsx --format esm,cjs --dts --external react --inject react-shim.js",
+    "dev": "tsup src/index.tsx --format esm,cjs --watch --dts --external react --inject react-shim.js",
     "lint": "TIMING=1 eslint src --fix",
     "clean": "rm -rf .turbo && rm -rf node_modules && rm -rf dist"
   },
