+++ examples/kitchen-sink/packages/ui/react-shim.js
@@ -1,0 +1,3 @@
+// Explanation for why this is required: https://esbuild.github.io/content-types/#auto-import-for-jsx
+import React from 'react'
+export { React }
