+++ examples/kitchen-sink/packages/ui/src/Button/index.tsx
@@ -1,0 +1,11 @@
+export const Button = ({ children, ...other }) => {
+  return (
+    <button 
+      target="_blank" 
+      rel="noreferrer" 
+      {...other}
+    >
+      {children}
+    </button>
+  );
+};
